videosearcher-backend/requirements.txt
fastapi==0.111.1
uvicorn==0.30.3
motor==3.5.1
pymongo==4.8.0
ffmpeg-python==0.2.0
opencv-python==4.10.0.84
torch==2.3.1
torchvision==0.18.1
requests==2.32.3
moviepy==1.0.3
pydantic~=2.8.2
numpy~=2.0.1
python-dotenv==0.19.1
pytesseract==0.3.10
pydantic-mongo==2.3.0
pillow~=10.4.0